Data.Map.Strict.NonEmptyMap

Description

Provides a strict implementation of a non-empty map.

This implementation is based on the implementation of Strict provided by the containers package, but provides an extra guarantee that the map contains at least one entry at all times.

Map type

data NonEmptyMap k v

A non-empty map from keys of type k to values of type v.

Construction

fromList :: Ord k => NonEmpty (k, v) -> NonEmptyMap k v

Builds a non-empty map from a list of key-value pairs.

If the list contains more than one value for the same key, the last value for the key is retained.

fromMap :: Map k v -> Maybe (NonEmptyMap k v)

Builds a non-empty map from the given map.

If the given map is empty, this function returns Nothing.

singleton :: Ord k => k -> v -> NonEmptyMap k v

Creates a map with a single element.

Deconstruction

toList :: NonEmptyMap k v -> NonEmpty (k, v)

Converts a non-empty map to a list of key-value pairs.

toMap :: Ord k => NonEmptyMap k v -> Map k v

Converts a non-empty map to an ordinary map.

Insertion

insert :: Ord k => k -> v -> NonEmptyMap k v -> NonEmptyMap k v

Inserts a new key and value in the map.

If the key is already present in the map, the associated value is replaced with the supplied value.

Deletion

delete :: Ord k => k -> NonEmptyMap k a -> Maybe (NonEmptyMap k a)

Deletes a key and its value from the map.

When the key is not a member of the map, the original map is returned.

This function returns Nothing if the delete operation reduces the number of elements to zero.

Lookup

lookup :: Ord k => k -> NonEmptyMap k v -> Maybe v

Looks up the value of a key in the map.

This function will return the corresponding value as '(Just value)', or Nothing if the key isn't in the map.

Combination

unionWith :: Ord k => (v -> v -> v) -> NonEmptyMap k v -> NonEmptyMap k v -> NonEmptyMap k v

Finds the union of two maps, with the given combining function.